+++ contracts/interfaces/IVerifierRegistry.sol
@@ -9,7 +9,7 @@ interface IVerifierRegistry {
     bool active,
     uint256 balance,
     uint256 shard,
-    bool enable
+    bool enabled
   );
 
   function uniqueNames(bytes32) external view returns (bool);
@@ -20,7 +20,7 @@ interface IVerifierRegistry {
   function isRegisteredVerifier(address) external view returns (bool);
 
   function updateActiveStatus(bool _active) external;
-  function updateEnableStatus(address _verifier, bool _enable) external;
+  function updateEnabledStatus(address _verifier, bool _enabled) external;
 
   function increaseShardBalance(address _verifier, uint256 _amount) external returns (bool);
   function decreaseShardBalance(address _verifier, uint256 _amount) external returns (bool);